pull

+++ resources/controller/usercontroller.py
@@ -1,7 +1,7 @@
 from model.user import UserModel
 from flask_restful import Resource,reqparse
 from flask import request as req
-from flask_jwt import jwt_required
+from flask_jwt_extended import *
 import json
 
 class UserController(Resource):
@@ -10,6 +10,7 @@ class UserController(Resource):
     parser.add_argument('name',required=True)
     parser.add_argument('email',required=True)
     parser.add_argument('phone',required=True)
+    parser.add_argument('isadmin',required=True,default=0)
     parser.add_argument('password',required=True)
 
     def post(self):
@@ -18,30 +19,43 @@ class UserController(Resource):
         user.save_to_db()
         return user.json()
     
-    @jwt_required()
+    @jwt_optional
     def get(self):
-        return list(map(lambda l:l.json(), UserModel.query.all())),201
+        islogin = get_jwt_identity()
+        users = list(map(lambda l:l.json(), UserModel.query.all()))
+        if islogin:
+            return user,201
+        return list(map(lambda x:x['name'],users))
 
 
 class AbsUserController(UserController):
     
-    @jwt_required()
+    @jwt_required
     def get(self,_id):
         users = UserModel.find_by_id(_id)
         if users:
             return users.json()
         return {"msg":'No user'},400
     
-    @jwt_required()
+    @jwt_required
     def delete(self,_id):
+        isadmin = get_jwt_claims()
+        print(isadmin )
+        if not isadmin['isadmin']:
+            return {'msg':'you dont have permision to delete this user'}
+            
         users = UserModel.find_by_id(_id)
         if users:
             users.delete_from_db()
             return users.json()
         return {"msg":'No user'},400
     
-    @jwt_required()
+    @jwt_required
     def put(self,_id):
+        isadmin = get_jwt_claims()
+        if isadmin['isadmin'] == 0:
+            return {'msg':'you dont have permision to delete this user'}
+            
         value = self.parser.parse_args()
         users = UserModel.find_by_id(_id)
         if users is not None:
